datatpe change for penalty kernel

+++ csrc/penaltyKernels.cu
@@ -31,7 +31,7 @@ namespace vllm
 constexpr float HALF_MAX = 65504.f;
 template <typename scalar_t>
 __global__ void batchApplyPenaltyKernel(scalar_t* input_logits, scalar_t* output_logits,
-    std::int32_t batch_size, std::int32_t vocab_size, std::int32_t max_seq_len,
+    std::int64_t batch_size, std::int64_t vocab_size, std::int64_t max_seq_len,
     std::int32_t* penalty_workspace,
     float const* temperatures,
     float const* repetition_penalties, float const* presence_penalties, float const* frequency_penalties,
@@ -165,7 +165,7 @@ __global__ void batchApplyPenaltyKernel(scalar_t* input_logits, scalar_t* output
 void batch_apply_penalty(
     torch::Tensor& input_logits,      
     torch::Tensor&  output_logits,
-    std::int32_t const batch_size, std::int32_t vocab_size, std::int32_t max_seq_len,
+    std::int64_t const batch_size, std::int64_t vocab_size, std::int64_t max_seq_len,
     torch::Tensor&  penalty_workspace,
     torch::Tensor& temperatures, torch::Tensor& repetition_penalties,
     torch::Tensor& presence_penalties, torch::Tensor& frequency_penalties,
